requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-analytics-data==0.17.0
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
mysql-connector-python==8.1.0
numpy==1.25.1
pandas==2.0.3
proto-plus==1.22.3
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
